Revert "Fix warnings"
This reverts commit dde4e1f6b0fb8e87789263eb00831bd9def1d829.

+++ src/mongoc/mongoc-iovec.h
@@ -32,7 +32,7 @@ BSON_BEGIN_DECLS
 #ifdef _WIN32
 typedef struct
 {
-   size_t  iov_len;
+   u_long  iov_len;
    char   *iov_base;
 } mongoc_iovec_t;
 #else
